Refactor views and windows file structure
