+++ app/components/views/__init__.py
@@ -17,5 +17,6 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
 
+from .ocr import FullScreenView
 from .preview import Preview
 from .viewContainer import ViewContainer
