+++ app/components/views/ocr/__init__.py
@@ -1,5 +1,5 @@
 """
-Cloe
+Cloe OCR View Components
 
 Copyright (C) `2021-2022` `<Alarcon Ace Belen>`
 
@@ -17,28 +17,4 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
 
-import sys
-
-from PyQt5.QtGui import QIcon
-from PyQt5.QtWidgets import QApplication
-
-from components.windows import SystemTray
-from utils.constants import APP_LOGO, APP_NAME, STYLESHEET_DEFAULT
-
-if __name__ == '__main__':
-
-    app = QApplication(sys.argv)
-    app.setApplicationName(APP_NAME)
-    app.setWindowIcon(QIcon(APP_LOGO))
-    app.setQuitOnLastWindowClosed(False)
-
-    widget = SystemTray()
-
-    styles = STYLESHEET_DEFAULT
-    with open(styles, 'r') as fh:
-        app.setStyleSheet(fh.read())
-    
-    widget.show()
-    widget.loadModel()
-    app.exec_()
-    sys.exit()
+from .fullscreen import FullScreenView
